Move brave_vpn_helper install methods to `brave/installer/win/util/`
An attempt to clean up layer violations

Also includes update to `GetVpnHelperServiceProfileDir`
(`brave_vpn_helper_utils.cc`) to use std::wstring to avoid additional
base::UTF8ToWide call.
