+++ browser/brave_vpn/sources.gni
@@ -30,9 +30,9 @@ if (enable_brave_vpn) {
     brave_browser_brave_vpn_deps += [
       "//brave/app:brave_generated_resources_grit",
       "//brave/browser:browser_process",
-      "//brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper:common",
       "//brave/components/brave_vpn/common/win",
       "//brave/components/brave_vpn/common/wireguard/win",
+      "//brave/installer/win/util:vpn_utils",
       "//chrome/common:constants",
       "//components/prefs",
       "//net",
