+++ chromium_src/chrome/installer/setup/sources.gni
@@ -16,8 +16,8 @@ if (is_win) {
   if (enable_brave_vpn) {
     brave_chromium_src_chrome_installer_setup_deps += [
       "//brave/components/brave_vpn/browser/connection/ikev2/win:ras_utils",
-      "//brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper:common",
       "//brave/components/brave_vpn/common/wireguard/win",
+      "//brave/installer/win/util:vpn_utils",
     ]
   }
   if (is_official_build) {
