+++ chromium_src/chrome/installer/setup/uninstall.cc
@@ -10,11 +10,11 @@
 #include "brave/installer/util/brave_shell_util.h"
 
 #if BUILDFLAG(ENABLE_BRAVE_VPN)
-#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/brave_vpn_helper_constants.h"
-#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/service_details.h"
 #include "brave/components/brave_vpn/browser/connection/ikev2/win/ras_utils.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
+#include "brave/installer/win/util/brave_vpn_helper_constants.h"
+#include "brave/installer/win/util/brave_vpn_helper_utils.h"
 #endif
 #define UninstallProduct UninstallProduct_ChromiumImpl
