+++ components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/vpn_dns_handler.cc
@@ -11,12 +11,12 @@
 #include "base/strings/utf_string_conversions.h"
 #include "base/time/time.h"
 #include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/brave_vpn_dns_delegate.h"
-#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/brave_vpn_helper_constants.h"
-#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/service_details.h"
 #include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/vpn_utils.h"
 #include "brave/components/brave_vpn/browser/connection/ikev2/win/ras_utils.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_commands.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
+#include "brave/installer/win/util/brave_vpn_helper_constants.h"
+#include "brave/installer/win/util/brave_vpn_helper_utils.h"
 
 namespace brave_vpn {
