+++ components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/vpn_utils.h
@@ -23,11 +23,6 @@ HANDLE OpenWpmSession();
 bool CloseWpmSession(HANDLE engine);
 // Subscribes for RAS connection notification of any os vpn entry.
 bool SubscribeRasConnectionNotification(HANDLE event_handle);
-// Install VPN Service
-bool InstallBraveVPNHelperService();
-// Configure VPN Service autorestart.
-bool ConfigureServiceAutoRestart(const std::wstring& service_name,
-                                 const std::wstring& brave_vpn_entry);
 }  // namespace brave_vpn
 
 #endif  // BRAVE_COMPONENTS_BRAVE_VPN_BROWSER_CONNECTION_IKEV2_WIN_BRAVE_VPN_HELPER_VPN_UTILS_H_
