+++ components/brave_vpn/browser/connection/wireguard/win/BUILD.gn
@@ -19,14 +19,16 @@ source_set("win") {
   deps = [
     "//base",
     "//brave/components/brave_vpn/browser/api",
-    "//brave/components/brave_vpn/browser/connection:api",
-    "//brave/components/brave_vpn/browser/connection:common",
     "//brave/components/brave_vpn/browser/connection/wireguard",
     "//brave/components/brave_vpn/browser/connection/wireguard/credentials",
+    "//brave/components/brave_vpn/browser/connection:api",
+    "//brave/components/brave_vpn/browser/connection:common",
     "//brave/components/brave_vpn/common",
     "//brave/components/brave_vpn/common/buildflags",
     "//brave/components/brave_vpn/common/win",
     "//brave/components/brave_vpn/common/wireguard/win",
+    "//chrome/elevation_service:public_headers",
+    "//chrome/install_static:install_static_util",
     "//net",
     "//services/network/public/cpp",
   ]
