+++ components/brave_vpn/common/win/BUILD.gn
@@ -16,6 +16,7 @@ source_set("win") {
   visibility = [
     "//brave/browser/*",
     "//brave/components/brave_vpn/*",
+    "//brave/installer/win/util/*",
     "//chrome/browser/*",
   ]
   deps = [ "//base" ]
