+++ elevation_service/sources.gni
@@ -1,0 +1,16 @@
+# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//brave/components/brave_vpn/common/buildflags/buildflags.gni")
+
+brave_elevation_service_lib_deps =
+    [ "//brave/components/brave_vpn/common/buildflags" ]
+
+if (is_win && enable_brave_vpn) {
+  brave_elevation_service_lib_deps += [
+    "//brave/components/brave_vpn/common/wireguard/win",
+    "//brave/installer/win/util:vpn_utils",
+  ]
+}
