+++ installer/win/util/BUILD.gn
@@ -1,0 +1,24 @@
+# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+source_set("vpn_utils") {
+  sources = [
+    "brave_vpn_helper_constants.h",
+    "brave_vpn_helper_utils.cc",
+    "brave_vpn_helper_utils.h",
+  ]
+
+  deps = [
+    "//base",
+    "//brave/components/brave_vpn/common",
+    "//brave/components/brave_vpn/common/buildflags",
+    "//brave/components/brave_vpn/common/win",
+    "//brave/components/brave_vpn/common/wireguard/win:win",
+    "//chrome/elevation_service:public_headers",
+    "//chrome/install_static:install_static_util",
+    "//chrome/installer/util:with_no_strings",
+    "//components/version_info",
+  ]
+}
