+++ patches/chrome-elevation_service-elevator.cc.patch
@@ -1,98 +1,30 @@
 diff --git a/chrome/elevation_service/elevator.cc b/chrome/elevation_service/elevator.cc
-index 43526539839917601341cb33c5d7610bf4179f3b..e851901149d0f63822cbceaca5e8aee100facc1e 100644
+index 43526539839917601341cb33c5d7610bf4179f3b..a63d56b3b96fdc184b89ac6035ba0fa45fd515dd 100644
 --- a/chrome/elevation_service/elevator.cc
 +++ b/chrome/elevation_service/elevator.cc
-@@ -20,6 +20,20 @@
+@@ -17,6 +17,8 @@
+ #include "base/process/process.h"
+ #include "base/win/scoped_localalloc.h"
+ #include "base/win/win_util.h"
++#include "brave/components/brave_vpn/common/wireguard/win/wireguard_utils_win.h"
++#include "brave/installer/win/util/brave_vpn_helper_utils.h"
  #include "chrome/elevation_service/caller_validation.h"
  #include "chrome/elevation_service/elevated_recovery_impl.h"
  
-+// TODO(bsclifton): clean this up
-+#include "base/files/file_util.h"
-+#include "base/process/launch.h"
-+#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/brave_vpn_helper_constants.h"
-+#include "brave/components/brave_vpn/browser/connection/ikev2/win/brave_vpn_helper/service_details.h"
-+#include "brave/components/brave_vpn/common/buildflags/buildflags.h"
-+#include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
-+#include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
-+#include "brave/components/brave_vpn/common/wireguard/win/wireguard_utils_win.h"
-+#include "base/win/scoped_handle.h"
-+#include "base/win/windows_types.h"
-+#include "brave/components/brave_vpn/common/win/scoped_sc_handle.h"
-+#include "brave/components/brave_vpn/common/win/utils.h"
-+
- namespace elevation_service {
- 
- namespace {
-@@ -41,6 +55,55 @@ base::Process GetCallingProcess() {
-   return base::Process(calling_process_handle);
- }
- 
-+HRESULT InstallBraveVPNHelperServiceImpersonated() {
-+  HRESULT hr = ::CoImpersonateClient();
-+  if (FAILED(hr)) {
-+    return hr;
-+  }
-+
-+  base::ScopedClosureRunner revert_to_self(
-+      base::BindOnce([]() { ::CoRevertToSelf(); }));
-+
-+  auto executable_path = brave_vpn::GetBraveVpnHelperServicePath();
-+  base::CommandLine command_line(executable_path);
-+  command_line.AppendSwitch(brave_vpn::kBraveVpnHelperInstall);
-+
-+  base::LaunchOptions options = base::LaunchOptions();
-+  options.feedback_cursor_off = true;
-+  options.wait = true;
-+  base::Process proc = base::LaunchProcess(command_line, options);
-+  if (!proc.IsValid()) {
-+    return brave_vpn::HRESULTFromLastError();
-+  }
-+
-+  return S_OK;
-+}
-+
-+HRESULT InstallBraveWireGuardServiceImpersonated() {
-+  HRESULT hr = ::CoImpersonateClient();
-+  if (FAILED(hr)) {
-+    return hr;
-+  }
-+
-+  base::ScopedClosureRunner revert_to_self(
-+      base::BindOnce([]() { ::CoRevertToSelf(); }));
-+
-+  auto executable_path = brave_vpn::GetBraveVPNWireguardServiceExecutablePath();
-+  base::CommandLine command_line(executable_path);
-+  command_line.AppendSwitch(
-+      brave_vpn::kBraveVpnWireguardServiceInstallSwitchName);
-+
-+  base::LaunchOptions options = base::LaunchOptions();
-+  options.feedback_cursor_off = true;
-+  options.wait = true;
-+  base::Process proc = base::LaunchProcess(command_line, options);
-+  if (!proc.IsValid()) {
-+    return brave_vpn::HRESULTFromLastError();
-+  }
-+
-+  return S_OK;
-+}
-+
- }  // namespace
- 
- HRESULT Elevator::RunRecoveryCRXElevated(const wchar_t* crx_path,
-@@ -195,6 +258,24 @@ HRESULT Elevator::DecryptData(const BSTR ciphertext,
+@@ -195,6 +197,24 @@ HRESULT Elevator::DecryptData(const BSTR ciphertext,
    return S_OK;
  }
  
 +HRESULT Elevator::InstallVPNServices() {
 +  if (!brave_vpn::IsBraveVPNHelperServiceInstalled()) {
-+    HRESULT hr = InstallBraveVPNHelperServiceImpersonated();
++    HRESULT hr = brave_vpn::InstallBraveVPNHelperServiceImpersonated();
 +    if FAILED(hr) {
 +      return hr;
 +    }
 +  }
 +
 +  if (!brave_vpn::wireguard::IsWireguardServiceRegistered()) {
-+    HRESULT hr = InstallBraveWireGuardServiceImpersonated();
++    HRESULT hr = brave_vpn::InstallBraveWireGuardServiceImpersonated();
 +    if FAILED(hr) {
 +      return hr;
 +    }
